Catch specific exception to check state type expected by der_der_cost
Operator has a custom excpetion that is raised when a der_der_cost
method that expects a vector is passed a dictionary. This is intended
to be caught by State.do_uncertainty. Previously this caught any
exception, hiding bugs during development. ValueError is more
specific but this can still be thrown elsewhere in der_der_cost.