calculates hessian diagonal for neighbouring parameters
Calculates the first off diagonal. This only produces output where
neighbouring pairs of parameters in the ordered dictionary are
constant or variable (if all pairs contain a fixed variable then the
output is empty). Needs to be extended to all pairs (whether
neighbours or not) and extended to remove null pairings.