Use NetCDF4 for storage
These changes have started the enabling of NetCDF4 for storage of
the state solution and associated uncertainties. This requires a
number of changes, namely specifying metadata and so on, which has
started to be done in the PROSAIL part of the code, but is not
complete. The documentation/docstrings also need updating, and
probably more sensible variable names should be used. At some stage,
I need to add saving the Hessian (code is there, but not yet
included). Note that the code for observations and what not is
currently in a different repo, but will get eventually merged, by
what time eoldas_ng will be an untamed beast with a million useless
dependencies.