State grid on spatial regularisation Hessian
This change adds support for the state grid in the spatial
regularisation operator. This has lead to changes on how the different
submatrices are built up (they are built up explicitly), which might be
suboptimal in terms of performance, but at least we know that it works.
I have tested the code with some data, and it apparently performs well.

A couple of things that need to be borne in mind is that the masked bits
now appear as NaNs, which is fine, but might need to be documented.
Also, the propagation from one date to the next needs to be considered
in terms of e.g. matrix sizes and so on (although I think it is OK as it
stands, it just needs testing).